Move everything quiz related over to its own folder
For OOP and everything else, quiz will be on its own world but only need to communicate with `Logic` and `Session`
